+++ CHANGELOG.md
@@ -4,10 +4,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.6.5] (Unreleased)
+## [0.6.5]
 ### Refactored
-- Remove `styled-atlaskit` folder and move its components up by one level
-- Remove `Form` folder and move its components up by one level
+- Remove `styled-atlaskit`, `Form`, `Loading`, `Navigation`, `Page`, and `Wrapper` folder and move its components up by one level
 
 ## [0.6.4]
 ### Fixed